requirements.txt
python==3.8.17
pytorch==2.0.1+cu117
numpy==1.25.2
scipy==1.14.1
scikit-learn==1.5.2
pandas==2.0.3
dgl==1.1.2+cu116